+++ scripts/psrflux/measure_flux_eda2.sh
@@ -20,12 +20,35 @@ if [[ -n "$3" && "$3" != "-" ]]; then
    pulse_end_bin=$3
 fi
 
+options="" # e.g. --add_sun --sun_ra=12.5822 --sun_dec=5.3932
+if [[ -n "$4" && "$4" != "-" ]]; then
+   options="$4"
+fi
+
+fitted_func="pulse_gauss_only"
+if [[ -n "$5" && "$5" != "-" ]]; then
+   fitted_func="$5"
+fi
+
+noise_start=0.00
+if [[ -n "$6" && "$6" != "-" ]]; then
+   noise_start=$6
+fi
+
+noise_end=1.00
+if [[ -n "$7" && "$7" != "-" ]]; then
+   noise_end=$7
+fi
+
 echo "#########################################"
 echo "PARAMETERS:"
 echo "#########################################"
 echo "ar_file = $ar_file"
 echo "pulse_start_bin = $pulse_start_bin"
 echo "pulse_end_bin = $pulse_end_bin"
+echo "options = $options"
+echo "fitted_func = $fitted_func"
+echo "Noise range : $noise_start - $noise_end"
 echo "#########################################"
 
 
@@ -77,7 +100,7 @@ python ~/github/station_beam/python/eda_sensitivity.py \
          --size=512 \
          --trcv_type=trcv_eda2 \
          -p None \
-         -m analytic \
+         -m analytic ${options} \
          > ${sens_file} 2>&1
 
 lastline=$(grep 'Stokes I images' ${sens_file})
@@ -106,7 +129,15 @@ python ${fp}/flux_calc.py $std_vcs $std_sim $bins $off_pts $period $prof_sum
 
 
 echo "MSOK's calculation in root"
-echo "plot_psr_profile.C(\"${pulse_profile_file}\",true,\"pulse_gauss\",0,1.0,$std_sim)"
+echo "plot_psr_profile.C(\"${pulse_profile_file}\",true,\"${fitted_func}\",0,1.0,$std_sim)"
 echo "WARNING : modify noise range manually !!!"
-cp ~/github/station_beam/scripts/psrflux/plot_psr_profile.C .
-root -l -q -b "plot_psr_profile.C(\"${pulse_profile_file}\",true,\"pulse_gauss\",0,1.0,$std_sim)"
+
+if [[ ! -s plot_psr_profile.C ]]; then
+   echo "cp ~/github/station_beam/scripts/psrflux/plot_psr_profile.C ."
+   cp ~/github/station_beam/scripts/psrflux/plot_psr_profile.C .
+else
+   echo "INFO : file plot_psr_profile.C found - no need to overwrite"
+fi   
+
+
+root -l -q -b "plot_psr_profile.C(\"${pulse_profile_file}\",true,\"${fitted_func}\",${noise_start},${noise_end},$std_sim)"
